Moved the io_data_buffer class.
